split expression and fixed tests
