+++ datasource/src/test/scala/quasar/physical/mongo/interpreter/WrapSpec.scala
@@ -21,25 +21,25 @@ import slamdata.Predef._
 import org.specs2.mutable.Specification
 import org.specs2.ScalaCheck
 
-import quasar.physical.mongo.{Version, Aggregator, MongoExpression => E}, E.ProjectionStep.Field
+import quasar.physical.mongo.Version
+import quasar.physical.mongo.expression._
 
 class WrapSpec extends Specification with ScalaCheck {
   "wrap interpreter produces two projects" >> prop { (wrapString: String, version: Version, rootKey: String) =>
     ((wrapString.length > 0) && (wrapString != "\u0000")) ==> {
-      val interpretation = Wrap(rootKey, version, Field(wrapString))
+      val interpretation = Wrap.apply0(rootKey, Field(wrapString))
       interpretation must beLike {
-        case Aggregator.project(E.Object(firstPair)) :: Aggregator.project(E.Object(secondPair)) :: List() =>
-          (secondPair._1 === rootKey) and
-          (secondPair._2 === E.key(firstPair._1)) and
-          (firstPair._2 === E.Object(wrapString -> E.key(rootKey)))
+        case Pipeline.$project(prj1) :: Pipeline.$project(prj2) :: List() =>
+          (prj1.get(rootKey concat "_wrap") === Some(O.obj(Map(wrapString -> O.key(rootKey))))) and
+          (prj2.get(rootKey) === Some(O.key(rootKey concat "_wrap")))
       }
     }
   }
   "wrap example" >> {
-    val actual = Wrap("root", Version.zero, Field("wrapper"))
+    val actual = Wrap.apply0("root", Field("wrapper"))
     val expected = List(
-      Aggregator.project(E.Object("root_wrap" -> E.Object("wrapper" -> E.key("root")))),
-      Aggregator.project(E.Object("root" -> E.key("root_wrap"))))
+      Pipeline.$project(Map("root_wrap" -> O.obj(Map("wrapper" -> O.key("root"))))),
+      Pipeline.$project(Map("root" -> O.key("root_wrap"))))
     actual === expected
   }
 }
